Batch update scores in the admin panel
https://github.com/EFForg/check-your-reps/issues/115
Future improvemements: currently, when you try to batch update
positions, ActiveAdmin shows you a dialog so you can choose what to
update the positions to.  That is great, but the dialog says "Are you
sure?" and it would be better if it said "To what shall I update
them?" or something.