Remove default position value from Scores
All congress members should have a score, so that admins can easily
update the score position. However, we should not show scores until
their score has been manually set by an admin. The easiest way to get
this effect is to make sure every congress member has a score, to
initialize scores with a nil position, and to hide scores without
positions from the list.

Whew.

So this is the first part of that: remove the position default.